pod event handling 0.1

+++ machine_manager/main.go
@@ -8,7 +8,6 @@ import (
 	pb "irio/linter_proto"
 	"log"
 	"net"
-	"os"
 	"strings"
 	"sync"
 	"time"
@@ -18,169 +17,248 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	apiv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
+
+	//appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/builder"
+	ctrlClient "sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/predicate"
 )
 
-type language = string
-type version = string
-type address = string
-type port = int32
+const (
+	ADD_WORKER PodEventType = iota
+	ADD_LB     PodEventType = iota
+	REMOVE_POD PodEventType = iota
+)
 
-type MachineInfo struct {
-	Address address
-	Port    port
-}
-type MachineInfoSet map[MachineInfo]bool
+const (
+	LOAD_BALANCER PodRole = iota
+	LINTER        PodRole = iota
+)
 
-func (mis MachineInfoSet) MarshalJSON() (data []byte, err error) {
-	machines := make([]MachineInfo, 0, len(mis))
-	for k := range mis {
-		machines = append(machines, k)
-	}
-	return json.Marshal(machines)
-}
+type PodRole int
+type PodEventType int
+type Language string
+type Version string
+type Address string
+type Port int32
 
-func (mis *MachineInfoSet) UnmarshalJSON(data []byte) error {
-	*mis = make(MachineInfoSet)
-
-	var machines []MachineInfo
-	err := json.Unmarshal(data, &machines)
-	if err != nil {
-		log.Println("error unmarshaling JSON: " + err.Error())
-		return err
-	}
-
-	for _, machine := range machines {
-		(*mis)[machine] = true
-	}
-	return nil
+type MachineManagerServer struct {
+	pb.UnimplementedMachineManagerServer
+	State      MachineManagerState
+	kubernetes kubernetes.Clientset
+	mut        sync.Mutex
+	eventChan  chan *PodEvent
 }
 
-type LBWorkersInfo struct {
-	Machines MachineInfoSet
-}
-type WorkersInfo struct {
-	Machines map[language]map[version]MachineInfoSet
-}
-
-func (lb *LBWorkersInfo) addMachine(machine MachineInfo) {
-	lb.Machines[machine] = true
+type MachineManagerState struct {
+	Persistent    MachineManagerPersistentState
+	LoadBalancers map[types.NamespacedName]LBInfo
+	Linters       map[types.NamespacedName]WorkerInfo
 }
 
-func (lb *LBWorkersInfo) removeMachine(machine MachineInfo) {
-	delete(lb.Machines, machine)
+type MachineManagerPersistentState struct {
+	Weights []*pb.Weight
 }
 
-func (w *WorkersInfo) getMachines() []*pb.Worker {
-	workers := []*pb.Worker{}
-	for lang, verMap := range w.Machines {
-		for ver, machMap := range verMap {
-			for mach := range machMap {
-				workers = append(workers, &pb.Worker{
-					Address: mach.Address,
-					Port:    mach.Port,
-					Attrs: &pb.LinterAttributes{
-						Language: lang,
-						Version:  ver,
-					},
-				})
+type WorkerInfo struct {
+	address Address
+	lang    Language
+	ver     Version
+}
+
+type LBInfo struct {
+	conn        grpc.ClientConn
+	cancelCtx   context.CancelFunc
+	commandChan chan<- bool
+}
+
+type PodEvent struct {
+	Type     PodEventType
+	Name     *types.NamespacedName
+	Addr     *Address
+	Version  *Version
+	Language *Language
+}
+
+type podInfo struct {
+	address string
+}
+
+type PodReconciler struct {
+	ctrlClient.Client
+	mut    sync.Mutex
+	pods   map[string]podInfo
+	events chan<- *PodEvent
+}
+
+const (
+	markerLabel                         = "xD"
+	requestTimeout                      = 3 * time.Second
+	machine_manager_config_map          = "machine-manager-config-map"
+	machine_manager_config_map_data_key = "json"
+	load_balancer_port                  = 33333
+	linter_port                         = 55555
+	num_retries                         = 3
+)
+
+var (
+	listen_addr = flag.String("address", ":2137", "The Admin CLI Listen address (with port)")
+	grpc_opts   = []grpc.DialOption{grpc.WithBlock(), grpc.WithInsecure()}
+)
+
+func (a *PodReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+	pod := &corev1.Pod{}
+	err := a.Get(ctx, req.NamespacedName, pod)
+	deleted := false
+	created := false
+	info := podInfo{}
+	ok := true
+	if apierrors.IsNotFound(err) {
+		a.mut.Lock()
+		if info, ok = a.pods[req.NamespacedName.String()]; ok {
+			delete(a.pods, req.NamespacedName.String())
+			deleted = true
+
+			a.events <- &PodEvent{
+				Type: REMOVE_POD,
+				Name: &req.NamespacedName,
 			}
 		}
+		a.mut.Unlock()
+	} else if err != nil {
+		fmt.Printf("Error processing event for %s: %s\n", req.NamespacedName, err)
+		return ctrl.Result{}, err
+	} else {
+		if pod.Status.Phase == corev1.PodRunning {
+			a.mut.Lock()
+			if _, ok = a.pods[req.NamespacedName.String()]; !ok {
+				address := pod.Status.PodIP
+				a.pods[req.NamespacedName.String()] = podInfo{
+					address: address,
+				}
+
+				role := LOAD_BALANCER
+				if pod.Labels[markerLabel] == "linter" {
+					role = LINTER
+				}
+				event := PodEvent{
+					Type: ADD_LB,
+					Name: &req.NamespacedName,
+					Addr: (*Address)(&address),
+				}
+				if role == LOAD_BALANCER {
+					event.Type = ADD_LB
+				} else {
+					event.Type = ADD_WORKER
+					ver := Version(pod.Labels["version"])
+					lang := Language(pod.Labels["language"])
+					event.Version = &ver
+					event.Language = &lang
+				}
+				a.events <- &event
+
+				info = a.pods[req.NamespacedName.String()]
+				created = true
+			}
+			a.mut.Unlock()
+		}
+	}
+
+	if created {
+		log.Printf("Created: %s %v\n", req.NamespacedName, info)
+	} else if deleted {
+		log.Printf("Deleted: %s %v\n", req.NamespacedName, info)
+	} else {
+		log.Printf("Changed: %s\n", req.NamespacedName)
+	}
+
+	return ctrl.Result{}, nil
+}
+
+func (mms *MachineManagerState) getWorkers() []*pb.Worker {
+	workers := []*pb.Worker{}
+	for _, worker := range mms.Linters {
+		workers = append(workers, &pb.Worker{
+			Address: string(worker.address),
+			Port:    linter_port,
+			Attrs: &pb.LinterAttributes{
+				Language: string(worker.lang),
+				Version:  string(worker.ver),
+			},
+		})
 	}
 	return workers
 }
 
-func (w *WorkersInfo) addMachine(lang language, ver version, machine MachineInfo) {
-	if _, exists := w.Machines[lang]; !exists {
-		w.Machines[lang] = make(map[version]MachineInfoSet)
+func (mms *MachineManagerState) hasMachinesFor(lang Language, ver Version) bool {
+	for _, worker := range mms.Linters {
+		if worker.lang == lang && worker.ver == ver {
+			return true
+		}
 	}
-	if _, exists := w.Machines[lang][ver]; !exists {
-		w.Machines[lang][ver] = make(MachineInfoSet)
-	}
-	w.Machines[lang][ver][machine] = true
+	return false
 }
 
-func (w *WorkersInfo) removeMachine(lang language, ver version, machine MachineInfo) {
-	map1, exists1 := w.Machines[lang]
-	if !exists1 {
-		return
-	}
-	map2, exists2 := w.Machines[lang][ver]
-	if !exists2 {
-		return
-	}
-	delete(map2, machine)
-	if len(map2) == 0 {
-		delete(map1, lang)
-	}
-	if len(map1) == 0 {
-		delete(w.Machines, lang)
-	}
-}
-
-func (w *WorkersInfo) removeMachinesForLinter(lang language, ver version) {
-	map_, exists1 := w.Machines[lang]
-	if !exists1 {
-		return
-	}
-	delete(map_, ver)
-	if len(map_) == 0 {
-		delete(w.Machines, lang)
-	}
-}
-
-var (
-	listen_addr = flag.String("address", "0.0.0.0:2137", "The Admin CLI Listen address (with port)")
-	grpc_opts   = []grpc.DialOption{grpc.WithBlock(), grpc.WithInsecure()}
-
-	dialTimeout                = 2 * time.Second
-	requestTimeout             = 10 * time.Second
-	etcd_addrs                 = []string{"localhost:2379"}
-	machine_manager_config_map = "machine-manager-config-map"
-	linter_port                = 33333
-)
-
-type machineManagerState struct {
-	Load_balancers LBWorkersInfo
-	Linters        WorkersInfo
-}
-
-type machineManagerServer struct {
-	pb.UnimplementedMachineManagerServer
-	state  machineManagerState
-	client kubernetes.Clientset
-	mut    sync.Mutex
-}
-
-func makeMachineManager() machineManagerServer {
-	l := log.New(os.Stderr, "", 0)
-	l.Println(os.Environ())
-	l.Println(os.Getenv("KUBERNETES_SERVICE_HOST"))
-	l.Println(os.Getenv("KUBERNETES_SERVICE_PORT"))
-
+func makeMachineManager() MachineManagerServer {
 	config, err := rest.InClusterConfig()
 	if err != nil {
 		panic(err.Error())
 	}
-	// creates the clientset
-	client, err := kubernetes.NewForConfig(config)
+	kubernetes, err := kubernetes.NewForConfig(config)
 	if err != nil {
 		panic(err.Error())
 	}
-	state := machineManagerState{
-		Load_balancers: LBWorkersInfo{Machines: make(MachineInfoSet)},
-		Linters:        WorkersInfo{Machines: make(map[string]map[string]MachineInfoSet)},
+	eventChan := make(chan *PodEvent)
+
+	manager, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{})
+	if err != nil {
+		panic("couldn't create manager: " + err.Error())
+	}
+
+	pred, err := predicate.LabelSelectorPredicate(metav1.LabelSelector{MatchExpressions: []metav1.LabelSelectorRequirement{{Key: markerLabel, Operator: metav1.LabelSelectorOpExists}}})
+	if err != nil {
+		panic("could not create predicate: " + err.Error())
+	}
+
+	err = ctrl.
+		NewControllerManagedBy(manager).
+		For(&corev1.Pod{}, builder.WithPredicates(pred)).
+		Complete(&PodReconciler{
+			Client: manager.GetClient(),
+			pods:   make(map[string]podInfo),
+			events: eventChan,
+		})
+	if err != nil {
+		panic("could not create controller: " + err.Error())
+	}
+
+	err = manager.Start(ctrl.SetupSignalHandler())
+	if err != nil {
+		panic("could not start manager: " + err.Error())
+	}
+
+	state := MachineManagerState{
+		Persistent: MachineManagerPersistentState{
+			Weights: []*pb.Weight{},
+		},
+		LoadBalancers: make(map[types.NamespacedName]LBInfo),
+		Linters:       make(map[types.NamespacedName]WorkerInfo),
 	}
 
 	ctx, cancelCtx := context.WithTimeout(context.Background(), requestTimeout)
 	defer cancelCtx()
 
-	get, err := client.CoreV1().ConfigMaps("default").Get(ctx, machine_manager_config_map, metav1.GetOptions{})
+	get, err := kubernetes.CoreV1().ConfigMaps(apiv1.NamespaceDefault).Get(ctx, machine_manager_config_map, metav1.GetOptions{})
 	if err != nil {
 		panic(err.Error())
 	}
-	jsonData := get.Data["json"]
+	jsonData := get.Data[machine_manager_config_map_data_key]
 	if len(jsonData) != 0 {
 		err = json.Unmarshal([]byte(jsonData), &state)
 		if err != nil {
@@ -188,53 +266,27 @@ func makeMachineManager() machineManagerServer {
 		}
 	}
 
-	///////////////////////////////////////////////////////////////////////////
-	log.Printf("get: %v\n", get)
-	log.Println("jsonData: " + jsonData)
-	serializedState, err := json.Marshal(&state)
-	if err != nil {
-		panic(err.Error())
-	}
-	log.Println("serializedState: " + string(serializedState))
-	updatedMap := apiv1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "machine-manager-config-map",
-		},
-		Data: map[string]string{
-			"json": string(serializedState),
-		},
-	}
-	_, err = client.CoreV1().ConfigMaps("default").Update(ctx, &updatedMap, metav1.UpdateOptions{})
-	if err != nil {
-		panic(err.Error())
-	}
-	get, err = client.CoreV1().ConfigMaps("default").Get(ctx, machine_manager_config_map, metav1.GetOptions{})
-	if err != nil {
-		panic(err.Error())
-	}
-	log.Printf("get: %v\n", get)
-	///////////////////////////////////////////////////////////////////////////
-
-	return machineManagerServer{
-		state:  state,
-		client: *client,
-		mut:    sync.Mutex{},
+	// TODO: set sensible capacities of channels
+	return MachineManagerServer{
+		State:      state,
+		kubernetes: *kubernetes,
+		eventChan:  eventChan,
 	}
 }
 
 func int32Ptr(i int32) *int32 { return &i }
 
-func deploymentNameFromAttrs(lang language, ver version) string {
-	return fmt.Sprintf("linter-%s-%s", lang, strings.ReplaceAll(ver, ".", "-"))
+func deploymentNameFromAttrs(lang Language, ver Version) string {
+	return fmt.Sprintf("linter-%s-%s", string(lang), strings.ReplaceAll(string(ver), ".", "-"))
 }
 
-func deploymentFromLabels(lang language, ver version, imageUrl string) appsv1.Deployment {
+func deploymentFromLabels(lang Language, ver Version, imageUrl string) appsv1.Deployment {
 	linterName := deploymentNameFromAttrs(lang, ver)
 	labels := map[string]string{
-		"version":  ver,
-		"language": lang,
-		"app":      linterName,
-		"xd":      "linter",
+		"version":   string(ver),
+		"language":  string(lang),
+		"app":       linterName,
+		markerLabel: "linter",
 	}
 	return appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
@@ -271,12 +323,52 @@ func deploymentFromLabels(lang language, ver version, imageUrl string) appsv1.De
 	}
 }
 
-func (s *machineManagerServer) AppendLinter(ctx context.Context, req *pb.AppendLinterRequest) (*pb.LinterResponse, error) {
+func (s *MachineManagerPersistentState) linterWeight(lang Language, ver Version) float32 {
+	for _, w := range s.Weights {
+		if w.Attrs.Language == string(lang) && w.Attrs.Version == string(ver) {
+			return w.Weight
+		}
+	}
+	return 0
+}
+
+func (s *MachineManagerServer) pushLoadbalancerProportions() {
+	for _, lbInfo := range s.State.LoadBalancers {
+		select {
+		case lbInfo.commandChan <- true:
+		default:
+		}
+	}
+}
+
+func (s *MachineManagerServer) storeState() {
 	ctx, cancelCtx := context.WithTimeout(context.Background(), requestTimeout)
 	defer cancelCtx()
 
-	deploymentsClient := s.client.AppsV1().Deployments(apiv1.NamespaceDefault)
-	deployment := deploymentFromLabels(req.Attrs.Language, req.Attrs.Version, req.ImageUrl)
+	serializedState, err := json.Marshal(&s.State.Persistent.Weights)
+	if err != nil {
+		panic("error serializing state: " + err.Error())
+	}
+	updatedMap := apiv1.ConfigMap{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "machine-manager-config-map",
+		},
+		Data: map[string]string{
+			machine_manager_config_map_data_key: string(serializedState),
+		},
+	}
+	_, err = s.kubernetes.CoreV1().ConfigMaps(apiv1.NamespaceDefault).Update(ctx, &updatedMap, metav1.UpdateOptions{})
+	if err != nil {
+		panic("error saving state: " + err.Error())
+	}
+}
+
+func (s *MachineManagerServer) AppendLinter(ctx context.Context, req *pb.AppendLinterRequest) (*pb.LinterResponse, error) {
+	ctx, cancelCtx := context.WithTimeout(context.Background(), requestTimeout)
+	defer cancelCtx()
+
+	deploymentsClient := s.kubernetes.AppsV1().Deployments(apiv1.NamespaceDefault)
+	deployment := deploymentFromLabels(Language(req.Attrs.Language), Version(req.Attrs.Version), req.ImageUrl)
 	_, err := deploymentsClient.Create(ctx, &deployment, metav1.CreateOptions{})
 	if err != nil {
 		log.Println("error appending linter: " + err.Error())
@@ -285,12 +377,21 @@ func (s *machineManagerServer) AppendLinter(ctx context.Context, req *pb.AppendL
 	return &pb.LinterResponse{Code: pb.LinterResponse_SUCCESS}, nil
 }
 
-func (s *machineManagerServer) RemoveLinter(ctx context.Context, req *pb.LinterAttributes) (*pb.LinterResponse, error) {
+func (s *MachineManagerServer) RemoveLinter(ctx context.Context, req *pb.LinterAttributes) (*pb.LinterResponse, error) {
 	ctx, cancelCtx := context.WithTimeout(context.Background(), requestTimeout)
 	defer cancelCtx()
 
-	deploymentsClient := s.client.AppsV1().Deployments(apiv1.NamespaceDefault)
-	deploymentName := deploymentNameFromAttrs(req.Language, req.Version)
+	lang := Language(req.Language)
+	ver := Version(req.Version)
+	deploymentsClient := s.kubernetes.AppsV1().Deployments(apiv1.NamespaceDefault)
+	deploymentName := deploymentNameFromAttrs(lang, ver)
+	s.mut.Lock()
+	linterWeight := s.State.Persistent.linterWeight(lang, ver)
+	s.mut.Unlock()
+	if linterWeight > 0 {
+		log.Printf("attempt to remove linter %s of nonzero weight %f\n", deploymentName, linterWeight)
+		return &pb.LinterResponse{Code: pb.LinterResponse_SUCCESS}, nil
+	}
 	err := deploymentsClient.Delete(ctx, deploymentName, metav1.DeleteOptions{})
 	if err != nil {
 		log.Println("error removing linter: " + err.Error())
@@ -299,26 +400,102 @@ func (s *machineManagerServer) RemoveLinter(ctx context.Context, req *pb.LinterA
 	return &pb.LinterResponse{Code: pb.LinterResponse_SUCCESS}, nil
 }
 
-func (s *machineManagerServer) SetProportions(ctx context.Context, req *pb.LoadBalancingProportions) (*pb.LinterResponse, error) {
-	s.mut.Lock()
-	defer s.mut.Unlock()
-	for machine := range s.state.Load_balancers.Machines {
-		peerAddr := fmt.Sprintf("%s:%d", machine.Address, machine.Port)
-		conn, err := grpc.Dial(peerAddr, grpc_opts...)
-		log.Printf("Connected to: %v", peerAddr)
-		if err != nil {
-			log.Fatalf("fail to dial: %v", err)
-			// TODO: better error handling?
+func (s *MachineManagerServer) SetProportions(ctx context.Context, req *pb.LoadBalancingProportions) (*pb.LinterResponse, error) {
+	for _, weight := range req.Weights {
+		lang := Language(weight.Attrs.Language)
+		ver := Version(weight.Attrs.Version)
+		s.mut.Lock()
+		hasMachines := s.State.hasMachinesFor(lang, ver)
+		s.mut.Unlock()
+		if !hasMachines {
+			log.Printf("attempt to set nonzero weight of %f for nonexistent linter %s \n", weight.Weight, deploymentNameFromAttrs(lang, ver))
+			return &pb.LinterResponse{Code: 1}, nil
 		}
-		defer conn.Close()
-		client := pb.NewLoadBalancerClient(conn)
-
-		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-		defer cancel()
-
-		client.SetConfig(ctx, &pb.SetConfigRequest{Workers: s.state.Linters.getMachines(), Weights: req.Weights})
 	}
+	s.mut.Lock()
+	s.State.Persistent.Weights = req.Weights
+	s.storeState()
+	s.mut.Unlock()
+	s.pushLoadbalancerProportions()
 	return &pb.LinterResponse{Code: pb.LinterResponse_SUCCESS}, nil
+}
+
+func (s *MachineManagerServer) handlePodEvents() {
+	for event := range s.eventChan {
+		switch event.Type {
+		case ADD_WORKER:
+			s.mut.Lock()
+			s.State.Linters[*event.Name] = WorkerInfo{
+				address: *event.Addr,
+				lang:    *event.Language,
+				ver:     *event.Version,
+			}
+			s.storeState()
+			s.mut.Unlock()
+			s.pushLoadbalancerProportions()
+		case ADD_LB:
+			cmdChan := make(chan bool, 1)
+			lbAddr := fmt.Sprintf("%s:%d", string(*event.Addr), load_balancer_port)
+
+			conn, err := grpc.Dial(lbAddr, grpc_opts...)
+			if err != nil {
+				log.Println("failed to dial %s: %s\n", lbAddr, err)
+			}
+			log.Printf("connected to: %v", lbAddr)
+			client := pb.NewLoadBalancerClient(conn)
+
+			ctx, cancelCtx := context.WithTimeout(context.Background(), requestTimeout)
+			go func() {
+				for _ = range cmdChan {
+					s.mut.Lock()
+					config := pb.SetConfigRequest{
+						Workers: s.State.getWorkers(),
+						Weights: s.State.Persistent.Weights,
+					}
+					s.mut.Unlock()
+					for i := 0; i < num_retries; i++ {
+						goCtx, cancelGoCtx := context.WithTimeout(ctx, requestTimeout)
+
+						_, err = client.SetConfig(goCtx, &config)
+						if err != nil {
+							log.Printf("failed to set config on a load balancer %s: %s", event.Name, err)
+						}
+						cancelGoCtx()
+					}
+					if err != nil {
+						log.Println("retries failed")
+					}
+				}
+			}()
+			s.mut.Lock()
+			s.State.LoadBalancers[*event.Name] = LBInfo{
+				conn:        *conn,
+				cancelCtx:   cancelCtx,
+				commandChan: cmdChan,
+			}
+			s.mut.Unlock()
+		case REMOVE_POD:
+			if lbInfo, isLb := s.State.LoadBalancers[*event.Name]; isLb {
+				delete(s.State.LoadBalancers, *event.Name)
+				lbInfo.cancelCtx()
+				lbInfo.conn.Close()
+				close(lbInfo.commandChan)
+			} else {
+				s.mut.Lock()
+				_, isWorker := s.State.Linters[*event.Name]
+				if isWorker {
+					delete(s.State.Linters, *event.Name)
+					s.storeState()
+				}
+				s.mut.Unlock()
+				if isWorker {
+					s.pushLoadbalancerProportions()
+				}
+			}
+		default:
+			log.Printf("wrong PodEvent type: %d", event.Type)
+		}
+	}
 }
 
 func main() {
@@ -333,5 +510,7 @@ func main() {
 	var server_opts []grpc.ServerOption
 	grpcServer := grpc.NewServer(server_opts...)
 	pb.RegisterMachineManagerServer(grpcServer, &machine_manager)
-	log.Fatal(grpcServer.Serve(lis))
+
+	go log.Fatal(grpcServer.Serve(lis))
+	machine_manager.handlePodEvents()
 }
